+++ src/clothes_shop/tests/categories/test_serializers.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 
 from clothes_shop.models.attributes import Brand, ClothesType, Size, Target
-from clothes_shop.serializers import (
+from clothes_shop.serializers.attributes_serializers import (
     BrandSerializer,
     ClothesTypeSerializer,
     SizeSerializer,
