+++ src/clothes_shop/tests/product/test_views.py
@@ -8,7 +8,7 @@ from rest_framework.test import APITestCase
 
 from clothes_shop.models.attributes import Brand, ClothesType, Size, Target
 from clothes_shop.models.product import Product
-from clothes_shop.serializers import ProductSerializer
+from clothes_shop.serializers.product_serializers import ProductSerializer
 from clothes_shop.services.stripe_service import StripeService
 
 stripe_service = StripeService()
