+++ src/clothes_shop/tests/review/test_views.py
@@ -10,7 +10,7 @@ from clothes_shop.models.attributes import Brand, ClothesType, Size, Target
 from clothes_shop.models.product import Product
 from clothes_shop.models.user import User
 from clothes_shop.models.user_interaction import Review
-from clothes_shop.serializers import ReviewSerializer
+from clothes_shop.serializers.user_interaction_serializers import ReviewSerializer
 
 logger = logging.getLogger(__name__)
