+++ src/clothes_shop/views/category_views.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from clothes_shop.models.attributes import Brand, ClothesType, Size, Target
-from clothes_shop.serializers import (
+from clothes_shop.serializers.attributes_serializers import (
     BrandSerializer,
     ClothesTypeSerializer,
     SizeSerializer,
