+++ src/clothes_shop/views/favorite_views.py
@@ -4,7 +4,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import generics
 
 from clothes_shop.models.user_interaction import Favorite
-from clothes_shop.serializers import FavoriteSerializer
+from clothes_shop.serializers.user_interaction_serializers import FavoriteSerializer
 
 logger = logging.getLogger(__name__)
