+++ src/clothes_shop/views/payment_views.py
@@ -4,7 +4,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import generics
 
 from clothes_shop.models.payment import Payment
-from clothes_shop.serializers import PaymentSerializer
+from clothes_shop.serializers.payment_serializers import PaymentSerializer
 
 logger = logging.getLogger(__name__)
