+++ src/clothes_shop/views/shipping_views.py
@@ -4,7 +4,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import generics
 
 from clothes_shop.models.shipping import Shipping
-from clothes_shop.serializers import ShippingSerializer
+from clothes_shop.serializers.shipping_serializers import ShippingSerializer
 
 logger = logging.getLogger(__name__)
